Quantization tool example bug fix (#133)
In ResNet50DataReader, it uses an onnx session to obtain the model input shape. However it passes a madeup model name to the onnx session, resulting in file not found error.

This change provide the original float model path to the data reader
